+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [0.3.14] - 2026-03-23
+
+### Other
+- Update .redup/cache/hash_cache.db
+- Update project/duplication.toon
+
 ## [0.3.13] - 2026-03-23
 
 ### Docs